refactor: reorganize files and functions
