+++ lib/plugin-config-resolution.ts
