+++ lib/rule-list.ts
@@ -1,4 +1,7 @@
-import { BEGIN_RULE_LIST_MARKER, END_RULE_LIST_MARKER } from './markers.js';
+import {
+  BEGIN_RULE_LIST_MARKER,
+  END_RULE_LIST_MARKER,
+} from './comment-markers.js';
 import {
   EMOJI_DEPRECATED,
   EMOJI_FIXABLE,
@@ -6,11 +9,11 @@ import {
   EMOJI_OPTIONS,
   EMOJI_REQUIRES_TYPE_CHECKING,
 } from './emojis.js';
-import { getConfigsForRule, findConfigEmoji } from './configs.js';
+import { getEmojisForConfigsSettingRuleToSeverity } from './plugin-configs.js';
 import { getColumns, COLUMN_HEADER } from './rule-list-columns.js';
 import { findSectionHeader } from './markdown.js';
 import { getPluginRoot } from './package-json.js';
-import { generateLegend } from './legend.js';
+import { generateLegend } from './rule-list-legend.js';
 import { relative } from 'node:path';
 import { COLUMN_TYPE, SEVERITY_TYPE } from './types.js';
 import { markdownTable } from 'markdown-table';
@@ -59,39 +62,6 @@ function getPropertyFromRule(
     result = result[part];
   }
   return result;
-}
-
-/**
- * Get the emojis for the configs that set a rule to a certain severity.
- */
-function getEmojisForConfigsSettingRuleToSeverity(
-  ruleName: string,
-  configsToRulesWithoutIgnored: ConfigsToRules,
-  pluginPrefix: string,
-  configEmojis: ConfigEmojis,
-  severityType: SEVERITY_TYPE
-) {
-  const configsOfThisSeverity = getConfigsForRule(
-    ruleName,
-    configsToRulesWithoutIgnored,
-    pluginPrefix,
-    severityType
-  );
-
-  const emojis: string[] = [];
-  for (const configName of configsOfThisSeverity) {
-    // Find the emoji for each config or otherwise use a badge that can be defined in markdown.
-    const emoji = findConfigEmoji(configEmojis, configName, {
-      fallback: 'badge',
-    });
-    /* istanbul ignore next -- this shouldn't happen */
-    if (typeof emoji !== 'string') {
-      throw new TypeError('Emoji will always be a string thanks to fallback');
-    }
-    emojis.push(emoji);
-  }
-
-  return emojis;
 }
 
 function getConfigurationColumnValueForRule(
