+++ test/lib/generate/option-check-test.ts
