+++ test/lib/generate/option-config-emoji-test.ts
