+++ test/lib/generate/option-rule-doc-notices-test.ts
