+++ test/lib/generate/option-rule-doc-section-test.ts
