+++ test/lib/generate/option-rule-doc-title-format-test.ts
