+++ test/lib/generate/option-rule-list-columns-test.ts
