+++ test/lib/generate/option-url-configs-test.ts
