+++ test/lib/generate/option-url-rule-doc-test.ts
