+++ test/lib/generate/rule-deprecation-test.ts
